Fix bug: precision loss from db index (long -> int)

+++ src/main/java/redis/clients/jedis/BinaryClient.java
@@ -73,7 +73,7 @@ public class BinaryClient extends Connection {
     this.password = password;
   }
   
-  public void setDb(long db) {
+  public void setDb(int db) {
     this.db = db;
   }
